chore(agent): Splitting Otlp Ingester code
